fix: address TTS settings review feedback
